third_party/WebKit/LayoutTests/fast/canvas-api/canvas-path-object.html

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<body>
<canvas id="canvas" width="200" height="200"></canvas>
<script>
test(function(t) {
    assert_equals(typeof document.getElementById, 'function');
    
    var path = new Path2D();
    assert_true(path instanceof Path2D);
    assert_equals(typeof path.closePath, 'function');
    assert_equals(typeof path.moveTo, 'function');
    assert_equals(typeof path.lineTo, 'function');
    assert_equals(typeof path.quadraticCurveTo, 'function');
    assert_equals(typeof path.bezierCurveTo, 'function');
    assert_equals(typeof path.arcTo, 'function');
    assert_equals(typeof path.arc, 'function');
    assert_equals(typeof path.rect, 'function');
    assert_equals(typeof path.ellipse, 'function');
}, "Ensure that the constructor for Path object and dependent functions exist.");
</script>
</body>
